Refactor compile command parsing
Refactor the compile command parsing to make it more aware of what each
of the compiler options means. Categorize each of the compiler options
as either for the preprocessor, C compiler, or C++ compiler.

Expand relative paths in the include path specifications to their
absolute equivalent.

Add Mocha unit test suites that cover all of the new files added to the
extension source code base.